mcp_server/pyproject.toml
[project]
name = "mcp_server"
version = "0.1.0"
description = "Amala Atlas MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopy>=2.4.1",
    "googlemaps>=4.10.0",
    "mcp[cli]>=1.14.0",
    "pip>=25.2",
    "python-dotenv>=1.1.1",
    "spacy>=3.8.7",
]
